+++ Runline/App/AppTab.swift
@@ -2,6 +2,7 @@ import SwiftUI
 
 enum AppTab: String, CaseIterable, Identifiable {
     case chats
+    case sdk
     case repositories
     case settings
 
@@ -11,6 +12,8 @@ enum AppTab: String, CaseIterable, Identifiable {
         switch self {
         case .chats:
             "Chats"
+        case .sdk:
+            "SDK"
         case .repositories:
             "Repositories"
         case .settings:
@@ -22,6 +25,8 @@ enum AppTab: String, CaseIterable, Identifiable {
         switch self {
         case .chats:
             "message"
+        case .sdk:
+            "terminal"
         case .repositories:
             "folder"
         case .settings: